fix(@ngtools/webpack): check for compiler-cli on plugin constructor
Fix #7904
